Rework to use recursive search
Directory.GetFile() has a *really* dumb problem where if you hit a directory that you don't have read access to, it completely errors out and you lose all data collected so far. This solution works a bit better, but since we have to walk all drives, I am only going through it 1 time and adding files with the right extenstions to a list.