Adding ETWEventSubscription
This assembly is analagous to WMI event subscriptions in that it waits for a predetermined event to occur and then executes the programmed function (in this case, `DoEvil()`). Instead of WMI, this tool leverages Event Tracing for Windows (ETW) sessions to monitor for either user logon events through the `Microsoft-Windows-Winlogon` provider or process creation events using the `Microsoft-Windows-Kernel-Process` kernel provider. This means that we are not going to get picked up on Sysmon events 19-21.