Adding unicode/non-printable service tool
We can create services using Unicode and/or non-printable characters that function as intended but don't show in Task Manager, services.msc, sc.exe, etc, making them difficult for defenders to spot and remove. This tool allows us to check for and remove them. This is only the defensive portion. The offensive POC will be added shortly.